+++ packages/delta-protocol-common/CHANGELOG.md
@@ -6,6 +6,8 @@
   * Also remove all cross-translation code relating to those.
 * Propagate the change to the `ResolveInfoDeducer` type from the `core` package.
 * Make all message parameters that are declared with **`?`** postfix optional, i.e.: the corresponding key doesn’t need to be present in the JSON.
+* Adjust the `Reference{Added|Changed|Deleted}Event` and `{Add|Change|Delete}ReferenceCommand` types, replacing “target” with “reference”, and propagate these changes to the entire codebase.
+  *Note* that this is a **breaking change**, due to a [change to the protocol specification](https://github.com/LionWeb-io/specification/issues/414), without providing backward compatibility e.g. by only deprecating those classes.
 
 
 ## 0.7.2